# docs/api/openapi-stage1.yaml
openapi: 3.0.3
info:
  title: KPK Platform API (Stage 1)
  version: 0.1.0
servers:
  - url: https://example.org/api
paths:
...
  /search:
    post:
      summary: Поиск чанков с учётом даты и цитат
      description: |
        К запросу применяется та же схема эмбеддингов, что и при индексировании. Параметр ``as_of``
        фильтрует результаты по ``valid_from`` / ``valid_to``. Каждый элемент выдачи содержит список цитат
        с указанием норматива и координат чанка.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string, description: 'Поисковый запрос (до 4000 символов)' }
                as_of: { type: string, format: date, description: 'Фильтр по дате, ISO 8601' }
                top_k: { type: integer, default: 5 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SearchHit' }
  /graph:
    get:
      summary: Подграф Neo4j вокруг узла
      description: |
        Gateway инициирует пересборку графа (Postgres → Neo4j), затем возвращает подграф глубиной ``depth``
        вокруг ``node_id``.
      parameters:
        - in: query
          name: node_id
          required: true
          schema: { type: string }
        - in: query
          name: depth
          schema: { type: integer, minimum: 1, maximum: 5, default: 2 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GraphResponse' }
...
components:
  schemas:
...
    SearchCitation:
      type: object
      properties:
        norm_unit_id: { type: string, nullable: true }
        norm_ref: { type: string, nullable: true }
        chunk:
          type: object
          properties:
            chunk_id: { type: string }
            page: { type: integer, nullable: true }
            slide: { type: integer, nullable: true }
            timecode: { type: string, nullable: true }
    SearchHit:
      type: object
      properties:
        score: { type: number }
        chunk_id: { type: string }
        source_id: { type: string, nullable: true }
        norm_ref: { type: string, nullable: true }
        valid_from: { type: string, nullable: true }
        valid_to: { type: string, nullable: true }
        citations:
          type: array
          items: { $ref: '#/components/schemas/SearchCitation' }
    GraphNode:
      type: object
      properties:
        id: { type: string }
        labels:
          type: array
          items: { type: string }
        properties:
          type: object
    GraphEdge:
      type: object
      properties:
        type: { type: string }
        from:
          type: object
          properties:
            id: { type: string }
            labels:
              type: array
              items: { type: string }
        to:
          type: object
          properties:
            id: { type: string }
            labels:
              type: array
              items: { type: string }
        properties:
          type: object
    GraphResponse:
      type: object
      properties:
        nodes:
          type: array
          items: { $ref: '#/components/schemas/GraphNode' }
        edges:
          type: array
          items: { $ref: '#/components/schemas/GraphEdge' }
...